+++ 4-period/java-vendingmachine-precourse-main/src/main/java/vendingmachine/domain/money/Coin.java
@@ -1,4 +1,4 @@
-package vendingmachine;
+package vendingmachine.domain.money;
 
 public enum Coin {
     COIN_500(500),
