+++ 4-period/java-vendingmachine-precourse-main/src/main/java/vendingmachine/domain/money/Money.java
@@ -1,4 +1,4 @@
-package vendingmachine;
+package vendingmachine.domain.money;
 
 import java.util.Objects;
